+++ spec/test_app/app/controllers/filters_controller.rb
@@ -2,6 +2,6 @@ class FiltersController < ApplicationController
   before_filter :get_npw
 
   def get_npw
-    @npw = NobelPrizeWinner.select('nobel_prize_winners.*, nobel_prizes.category, nobel_prizes.year').joins(:nobel_prizes).tableficate(params[:nobel_prize_winners])
+    @npw = NobelPrizeWinner.select('nobel_prize_winners.*, nobel_prizes.*').joins(:nobel_prizes).tableficate(params[:nobel_prize_winners])
   end
 end