+++ changelog.markdown
@@ -11,6 +11,7 @@
  * `column` now takes a `cell_attrs` option to provide attributes for every `td` in the column, a proc can be passed in place of a string for attributes to customize on a per cell basis
  * `column` now places all unrecognized options into the `attrs` attribute which is used on a `col` tag
  * adding `empty` to display a message when a table has no data
+ * input types are automatically figured out by label name and DB column type
 
 ## 0.2.1
  * fix table generator to provide `scope` the proper argument and correct documentation